정단호/2024-10-1/add Stock Fund API

+++ src/main/java/com/kb/finance/dto/Stock.java
@@ -13,6 +13,6 @@ public class Stock {
     private String stockName;
     private int stockPrice;
     private double priceChangeRate;
-    private int stockVolume;
+    private long stockVolume;
 
 }
